fix: first rotate camera will shake
